Rework the message structure
Commands are now a proper type on the Elm side, and serialized
in a more structured way (which means a bit more work parsing in
go). For server->client messages, the ID field is no longer shared
between message types (some of which didn't use it), resulting in
somewhat deeper nesting.